// contexts/storage/storage.type.ts
export type StorageQueryParams = {
    path: string; // This will be relative to 'public/' (e.g., "templateCover" instead of "public/templateCover")
    limit: number;
    offset: number;
    searchTerm?: string;
    fileType?: Graphql.FileType;
    sortField?: Graphql.FileSortField;
};

// Union type for storage items (files and folders)
export type StorageItem = Graphql.FileInfo | Graphql.DirectoryInfo;
// ...
export type ViewMode = "grid" | "list";
// ...
export type FileInfo = Graphql.FileInfo;
export type DirectoryInfo = Graphql.DirectoryInfo;